doc + subclass pbt + pb2_utils
